fix: Fix reports systems count and bar chart (#1862)
This PR replaces the bar chart on the reports table with the donut chart. It also fixes the counts which were previously incorrect.

The "total number of systems" considered in the main reports view do not take into account neither the reported or unsupported systems in the X of Y systems.

Though, the bar chart displayed on each of the report entries does take them into account (displaying the unsupported systems in yellow, and the never-reported systems in gray)

Steps to Reproduce:

have one unsupported system reporting
have 2 more additional systems associated to the reported policy
Check reports page

Co-authored-by: Michael Johnson <micjohns@newbeginnings.usersys.redhat.com>