fix(SystemsTable): RHICOMPL-3728 - Disable selection when selecting (#1948)
Description of problem
Selecting systems in the "Edit systems" modal for an OpenSCAP policy is broken. Clicking on the checkboxes next to each system doesn't reliably result in that checkbox being ticked.

How reproducible
Always.

Steps to Reproduce:
Create an OpenSCAP policy.
Navigate to the "systems" tab for the policy.
Click the "edit systems" button on that page.
Start clicking checkboxes next to systems.

Actual results
The selected checkboxes might be toggled, or they might not. They might toggle immediately, or some seconds later, and they might toggle zero times, or once, or twice.

Expected results
The selected checkbox toggles immediately, and exactly once.

The issue is that we have to update the "Rules" tab in the "Edit Systems" modal with each system selection. The bug happens when clicking a second system while the api call is still being made to update the "Rules" tab.

The fix is to disable system selections while the api call is being made.

Co-authored-by: Michael Johnson <micjohns@newbeginnings.usersys.redhat.com>